backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "API con integración de Hacker News y scraping de libros"
authors = ["int_64 <carlosr301101gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
redis = "^5.0.1"
pydantic = "^2.4.2"
httpx = "^0.25.1"
beautifulsoup4 = "^4.12.2"
selenium = "^4.15.2"
webdriver-manager = "^4.0.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true